+++ apps/web/src/views/create-review/ui/RatingSelector.tsx
